ci: [ANDROAPP-4911] Move Injector.kt to di package
