+++ form/src/main/java/org/dhis2/form/ui/dialog/OptionSetDialog.kt
@@ -12,7 +12,7 @@ import androidx.fragment.app.DialogFragment
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.viewModels
 import com.google.android.material.composethemeadapter.MdcTheme
-import org.dhis2.form.Injector
+import org.dhis2.form.di.Injector
 import org.dhis2.form.model.DispatcherProvider
 import org.dhis2.form.model.FieldUiModel
 import org.dhis2.form.model.OptionSetDialogViewModel
